requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
TTS>=0.8.0
torch>=2.0.0
python-dotenv>=0.19.0
email-validator>=1.1.3